+++ src/Pages/Translate/TranslateFrom.jsx
@@ -14,6 +14,8 @@ import { pdfjs } from "react-pdf";
 import { AuthContext } from "../../Security/AuthProvider"
 import Feedback from "../../Components/Feedback/Feedback";
 import MyRating from "../../Components/Rating/MyRating";
+import BG from "../../Components/Features/BG";
+import Animation from "./Animation";
 
 
 function Translator() {
@@ -347,8 +349,9 @@ function Translator() {
   };
 
   return (
-    <div className="text-black bg-[#031321]  flex items-center justify-center">
-      <div className="bg-base-300 p-8 rounded-lg shadow-md w-4/5 my-28">
+    <div className="text-black bg-[#031321]  flex items-center justify-center ">
+      <Animation></Animation>
+      <div className="bg-base-300 p-8 rounded-lg shadow-md w-4/5 my-28 ">
         <h1 className="text-2xl text-center text-black font-bold mb-4">
           Translation Board
         </h1>
@@ -619,6 +622,7 @@ function Translator() {
           </div>
         </div>
       )}
+      
     </div>
   );
 }
